feat: dockerization
