+++ yolox-detection/requirements.txt
@@ -1,2 +1,15 @@
+tabulate==0.9.0
+torch==2.1.2
+numpy==1.26.2
+loguru==0.7.2
+torchvision==0.16.2
+opencv-python==4.8.1.78
+psutil==5.9.7
+tensorboard==2.15.1
+packaging==23.2
+pycocotools==2.0.7
+tqdm==4.66.1
+thop==0.1.1.post2209072238
+ninja==1.11.1.1
 picsellia==6.10.3
-
+onnx==1.15.0
